rename hooksnext to hooks
