+++ src/pages/namespace/Instances/List/Row.tsx
@@ -21,7 +21,7 @@ import TooltipCopyBadge from "~/design/TooltipCopyBadge";
 import { pages } from "~/util/router/pages";
 import { statusToBadgeVariant } from "../utils";
 import { useTranslation } from "react-i18next";
-import useUpdatedAt from "~/hooksNext/useUpdatedAt";
+import useUpdatedAt from "~/hooks/useUpdatedAt";
 
 const InstanceTableRow: FC<{
   instance: InstanceSchemaType;
